+++ packages/grid/src/config/TableConfig.ts
@@ -1,7 +1,9 @@
 export const tableBorderColor = "var(--lhds-color-ui-300)";
-export const tableBorderColorExpanded = "var(--lhds-color-blue-300)";
+export const tableBorderColorExpanded = "var(--lhds-color-blue-500)";
+export const tableBackgroundColorExpanded = "var(--lhds-color-blue-50)";
 export const tableBorder = `1px solid ${tableBorderColor}`;
-export const tableBorderExpanded = `1px solid ${tableBorderColorExpanded}`;
+export const tableBorderLeft = `2px solid transparent`;
+export const tableBorderLeftExpanded = `2px solid ${tableBorderColorExpanded}`;
 export const defaultTableRowHeight = "40px";
 export const defaultTableHeadRowHeight = "40px";
 export const smallTableRowWidth = "40px";